nevo/requirements.txt
appnope==0.1.0
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.9.10
django-appconf==1.0.2
django-compressor==2.1
django-cors-headers==1.2.2
django-extra-views==0.6.4
django-haystack==2.5.0
django-oscar==1.3
django-oscar-api==1.0.8
django-tables2==1.0.7
django-treebeard==4.0.1
django-widget-tweaks==1.4.1
djangorestframework==3.4.7
enum34==1.1.6
factory-boy==2.7.0
fake-factory==0.7.2
funcsigs==1.0.2
ipaddress==1.0.17
ipython==5.1.0
ipython-genutils==0.1.0
mock==2.0.0
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.1
phonenumbers==7.7.1
pickleshare==0.7.4
Pillow==3.4.1
prompt-toolkit==1.0.7
ptyprocess==0.5.1
purl==1.3
pycountry==1.20
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.7
rcssmin==1.0.6
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
sorl-thumbnail==12.4a1
traitlets==4.3.1
Unidecode==0.4.19
wcwidth==0.1.7
